Now registers itself as a sensor and figures out its IP-address. It will now try to automatically download certificates needed to communicate with the server..

+++ controller/src/main/java/se/cag/labs/rover/bean/Movement.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 public class Movement {
-    double left;
-    double right;
+    float left;
+    float right;
     boolean fire;
 }
